Smarter, faster, better all the things
Config generation is now done via python so we have a lot more control over what
and how things are generated. This change creates a more consist and
understandable application and does it in a simple, and uniform way.
Coincidentally, this change also lowers the CPU load on all relay nodes;
several performance bugs were addressed because the config was easier to
understand.

> There's basically a lot less Jinja...

When multiple nodes are present, CloudStream will now balance traffic from the
"main" node using an internal "private" network when present. This will ensure
there's a more equal egress network load on across all cluster members. The
load balancing is based on the stream bitrate. When the simulcast application
is generated it will map push destinations and will select targets with the
least saturation.

This is a major change and will ensure a far better streaming experience, fence
streams against failure, and lower the overall cost of operating a simulcast.

Signed-off-by: Kevin Carter <kecarter@redhat.com>